pyproject.toml
[project]
name = "octopus"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "httpx",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "ruff",
]

[tool.ruff]
target-version = "py314"
